Move ModuleBindings into a separate class
